+++ version_history.md
@@ -1,8 +1,7 @@
 # Version History
 
 ## 2.6.6
-- Added the ability to launch Marbleland levels directly from URLs
-- Added Marbleland leaderboard integration API endpoint.
+- Improved integration with Marbleland
 
 ## 2.6.5
 - Improved replay loading